[refactor]: rename coverage module class file names snake_case
