+++ demo/interpolate_symphonie_on_regular_grid.py
@@ -20,7 +20,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from spatialetl.coverage.TimeLevelCoverage import TimeLevelCoverage
+from spatialetl.coverage.time_level_coverage import TimeLevelCoverage
 from spatialetl.coverage.io.netcdf.symphonie.v293.SYMPHONIEReader import SYMPHONIEReader as CoverageReader
 from spatialetl.coverage.io.netcdf.DefaultWriter import DefaultWriter
 from spatialetl.utils.logger import logging
